fix: refactor account info queries
