+++ src/main/java/io/supertokens/storage/postgresql/queries/AccountInfoQueries.java
@@ -1,0 +1,106 @@
+/*
+ *    Copyright (c) 2025, VRAI Labs and/or its affiliates. All rights reserved.
+ *
+ *    This software is licensed under the Apache License, Version 2.0 (the
+ *    "License") as published by the Apache Software Foundation.
+ *
+ *    You may not use this file except in compliance with the License. You may
+ *    obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ *    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ *    License for the specific language governing permissions and limitations
+ *    under the License.
+ */
+
+package io.supertokens.storage.postgresql.queries;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+import io.supertokens.pluginInterface.ACCOUNT_INFO_TYPE;
+import io.supertokens.pluginInterface.multitenancy.TenantIdentifier;
+import static io.supertokens.storage.postgresql.QueryExecutorTemplate.update;
+import io.supertokens.storage.postgresql.Start;
+import io.supertokens.storage.postgresql.config.Config;
+import static io.supertokens.storage.postgresql.config.Config.getConfig;
+import io.supertokens.storage.postgresql.utils.Utils;
+
+public class AccountInfoQueries {
+    public static void addRecipeUserAccountInfo_Transaction(Start start, Connection sqlCon,
+                                                            TenantIdentifier tenantIdentifier, String userId,
+                                                            String recipeId, ACCOUNT_INFO_TYPE accountInfoType,
+                                                            String thirdPartyId, String thirdPartyUserId,
+                                                            String accountInfoValue)
+            throws SQLException {
+        String QUERY = "INSERT INTO " + getConfig(start).getRecipeUserTenantsTable()
+                + "(app_id, recipe_user_id, tenant_id, recipe_id, account_info_type, third_party_id, third_party_user_id, account_info_value)"
+                + " VALUES(?, ?, ?, ?, ?, ?, ?, ?)";
+
+        update(sqlCon, QUERY, pst -> {
+            pst.setString(1, tenantIdentifier.getAppId());
+            pst.setString(2, userId);
+            pst.setString(3, tenantIdentifier.getTenantId());
+            pst.setString(4, recipeId);
+            pst.setString(5, accountInfoType.toString());
+            pst.setString(6, thirdPartyId);
+            pst.setString(7, thirdPartyUserId);
+            pst.setString(8, accountInfoValue);
+        });
+    }
+
+    static String getQueryToCreateRecipeUserTenantsTable(Start start) {
+        String schema = Config.getConfig(start).getTableSchema();
+        String tableName = Config.getConfig(start).getRecipeUserTenantsTable();
+        // @formatter:off
+        return "CREATE TABLE IF NOT EXISTS " + tableName + " ("
+                + "app_id VARCHAR(64) NOT NULL,"
+                + "recipe_user_id CHAR(36) NOT NULL,"
+                + "tenant_id VARCHAR(64) NOT NULL,"
+                + "recipe_id VARCHAR(128) NOT NULL,"
+                + "account_info_type VARCHAR(8) NOT NULL,"
+                + "account_info_value TEXT NOT NULL,"
+                + "third_party_id VARCHAR(28),"
+                + "third_party_user_id VARCHAR(256),"
+                + "CONSTRAINT " + Utils.getConstraintName(schema, tableName, null, "pkey")
+                + " PRIMARY KEY (app_id, tenant_id, recipe_id, account_info_type, third_party_id, third_party_user_id, account_info_value),"
+                + "CONSTRAINT " + Utils.getConstraintName(schema, tableName, "tenant_id", "fkey")
+                + " FOREIGN KEY(app_id, tenant_id)"
+                + " REFERENCES " + Config.getConfig(start).getTenantsTable() + " (app_id, tenant_id) ON DELETE CASCADE"
+                + ");";
+        // @formatter:on
+    }
+
+    static String getQueryToCreatePrimaryUserTenantsTable(Start start) {
+        String schema = Config.getConfig(start).getTableSchema();
+        String tableName = Config.getConfig(start).getPrimaryUserTenantsTable();
+        // @formatter:off
+        return "CREATE TABLE IF NOT EXISTS " + tableName + " ("
+                + "app_id VARCHAR(64) NOT NULL,"
+                + "tenant_id VARCHAR(64) NOT NULL,"
+                + "account_info_type VARCHAR(8) NOT NULL,"
+                + "account_info_value TEXT NOT NULL,"
+                + "primary_user_id CHAR(36) NOT NULL,"
+                + "CONSTRAINT " + Utils.getConstraintName(schema, tableName, null, "pkey")
+                + " PRIMARY KEY (app_id, tenant_id, account_info_type, account_info_value)"
+                + ");";
+        // @formatter:on
+    }
+
+    static String getQueryToCreateTenantIndexForRecipeUserTenantsTable(Start start) {
+        return "CREATE INDEX IF NOT EXISTS idx_recipe_user_tenants_tenant ON "
+                + Config.getConfig(start).getRecipeUserTenantsTable() + "(app_id, tenant_id);";
+    }
+
+    static String getQueryToCreateAccountInfoIndexForRecipeUserTenantsTable(Start start) {
+        return "CREATE INDEX IF NOT EXISTS idx_recipe_user_tenants_account_info ON "
+                + Config.getConfig(start).getRecipeUserTenantsTable()
+                + "(app_id, tenant_id, account_info_type, third_party_id, account_info_value);";
+    }
+
+    static String getQueryToCreatePrimaryUserIndexForPrimaryUserTenantsTable(Start start) {
+        return "CREATE INDEX IF NOT EXISTS idx_primary_user_tenants_primary ON "
+                + Config.getConfig(start).getPrimaryUserTenantsTable() + "(app_id, primary_user_id);";
+    }
+}
